LIHADOOP-37013. Tensorflow execution type multiplexing between TF on Spark and TF on YARN
