+++ CONTRIBUTORS.md
@@ -40,6 +40,7 @@ The following were contributed by Anthony Hsu. Thanks, Anthony!
 * `LIHADOOP-16591: ligradle azkabanUpload fails with "ClassNotFoundException: org.apache.http.client.methods.HttpPost"`
 
 The following were contributed by Jonathan Hung. Thanks, Jonathan!
+* `LIHADOOP-37013. Tensorflow execution type multiplexing between TF on Spark and TF on YARN`
 * `Remove taskCommand in tensorflow dsl job`
 * `LIHADOOP-35318 Implement TensorFlow job in hadoop dsl`
 * `TOOLS-67569 Block Hadoop Client Teams from checking in Hadoop Jars in their MP`
