+++ VERSIONS.md
@@ -17,6 +17,10 @@ the License.
 Note that the LinkedIn build system occasionally requires that we skip a
 version bump, so you will see a few skipped version numbers in the list below.
 
+0.14.11
+
+* TensorFlow job multiplex between TF on Spark and TF on YARN
+
 0.14.10
 
 * DSL Trigger Implementation
