+++ gradle.properties
@@ -1,2 +1,2 @@
 org.gradle.daemon=true
-version=0.14.10
+version=0.14.11
