+++ hadoop-plugin-test/expectedJobs/jobs1/jobs1.job
@@ -1,3 +1,3 @@
 # This file generated from the Hadoop DSL. Do not edit by hand.
 type=noop
-dependencies=jobs1_job24
+dependencies=jobs1_job25
