+++ hadoop-plugin-test/expectedJobs/jobs1/jobs1_job24.job
@@ -4,7 +4,7 @@ dependencies=jobs1_job23
 am_memory=2048
 am_vcores=1
 archive=tensorflow-starter-kit.zip
-jar=tensorflow-on-yarn-0.0.1.jar
+executes=path/to/python/script.py
 num_ps=2
 num_workers=4
 ps_memory=2048
